+++ annotation.csv
@@ -1,51 +1,51 @@
 Absolute path,Relative path
-C:\Users\PC\Desktop\programis\ap_lab2\images\000001.jpg,000001.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000002.jpg,000002.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000003.jpg,000003.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000004.jpg,000004.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000005.jpg,000005.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000006.jpg,000006.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000007.jpg,000007.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000008.jpg,000008.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000009.jpg,000009.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000010.jpg,000010.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000011.jpg,000011.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000012.jpg,000012.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000013.jpg,000013.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000014.jpg,000014.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000015.jpg,000015.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000016.jpg,000016.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000017.png,000017.png
-C:\Users\PC\Desktop\programis\ap_lab2\images\000018.jpg,000018.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000019.jpg,000019.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000020.jpg,000020.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000021.jpg,000021.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000022.jpg,000022.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000023.png,000023.png
-C:\Users\PC\Desktop\programis\ap_lab2\images\000024.jpg,000024.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000025.jpg,000025.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000026.jpg,000026.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000027.jpg,000027.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000028.jpg,000028.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000029.jpg,000029.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000030.jpg,000030.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000031.jpg,000031.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000032.jpg,000032.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000033.jpg,000033.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000034.jpg,000034.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000035.jpg,000035.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000036.jpg,000036.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000037.jpg,000037.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000038.jpg,000038.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000039.jpg,000039.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000040.jpg,000040.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000041.jpg,000041.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000042.jpg,000042.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000043.jpg,000043.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000044.jpg,000044.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000045.jpg,000045.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000046.jpg,000046.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000047.jpg,000047.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000048.jpg,000048.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000049.jpg,000049.jpg
-C:\Users\PC\Desktop\programis\ap_lab2\images\000050.jpg,000050.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000001.jpg,images\000001.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000002.jpg,images\000002.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000003.jpg,images\000003.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000004.jpg,images\000004.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000005.jpg,images\000005.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000006.jpg,images\000006.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000007.jpg,images\000007.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000008.jpg,images\000008.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000009.jpg,images\000009.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000010.jpg,images\000010.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000011.jpg,images\000011.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000012.jpg,images\000012.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000013.jpg,images\000013.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000014.jpg,images\000014.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000015.jpg,images\000015.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000016.png,images\000016.png
+C:\Users\PC\Desktop\programis\ap_lab2\images\000017.png,images\000017.png
+C:\Users\PC\Desktop\programis\ap_lab2\images\000018.jpg,images\000018.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000019.jpg,images\000019.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000020.jpg,images\000020.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000021.jpg,images\000021.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000022.jpg,images\000022.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000023.jpg,images\000023.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000024.jpg,images\000024.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000025.png,images\000025.png
+C:\Users\PC\Desktop\programis\ap_lab2\images\000026.jpg,images\000026.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000027.jpg,images\000027.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000028.jpg,images\000028.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000029.jpg,images\000029.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000030.jpg,images\000030.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000031.jpg,images\000031.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000032.jpg,images\000032.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000033.jpg,images\000033.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000034.jpg,images\000034.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000035.jpg,images\000035.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000036.jpg,images\000036.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000037.jpg,images\000037.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000038.jpg,images\000038.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000039.jpg,images\000039.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000040.jpg,images\000040.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000041.jpg,images\000041.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000042.jpg,images\000042.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000043.jpg,images\000043.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000044.jpg,images\000044.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000045.jpg,images\000045.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000046.jpg,images\000046.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000047.jpg,images\000047.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000048.jpg,images\000048.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000049.jpg,images\000049.jpg
+C:\Users\PC\Desktop\programis\ap_lab2\images\000050.jpg,images\000050.jpg
