+++ make_annotation.py
@@ -1,6 +1,21 @@
 import csv
 import os
 
+def get_dir_files(dir_path: str) -> list:
+    """
+    Finding names of files in directory
+    :param dir_path: path to our directory
+    :return: list of file names 
+    """
+    try:
+        paths_to_img = []
+        for root, dirs, files in os.walk(dir_path):
+            for file in files:
+                paths_to_img.append(os.path.join(root, file))
+            return paths_to_img
+    except:
+        raise NotADirectoryError("Invalid directory")
+
 def make_annot(annot_path: str, files: list) -> None:
     """
     Making annotation 
@@ -14,5 +29,5 @@ def make_annot(annot_path: str, files: list) -> None:
             
         for img_file in files:
             abs_path = os.path.abspath(img_file)  
-            rel_path = os.path.relpath(img_file, cur_dir)  
+            rel_path = os.path.relpath(os.path.join(cur_dir, img_file)) 
             writer.writerow([abs_path, rel_path])
